Fixed `BuddyListNotification`

+++ Sources/MapleStory/Packets/BuddyListNotification.swift
@@ -40,6 +40,6 @@ public extension BuddyListNotification {
         
         internal let value0: UInt8
         
-        public let channel: UInt8
+        public let channel: Int32
     }
 }
